Implement masked rendering
The new method of accelerated rendering relies on a two-step process.
First, we calculate a mask which indicates the areas of high detail in
the scene. Then, we apply a random selection process, choosing
highlighted pixels more often than the others. Finally, we use an
inpainting algorithm to fill in the blanks.

Most of this works just fine, except the inpainting step. My algorithm
needs work; I think the key change is to weigh interpolation samples by
the distance between the pixels they come from. That should fix the
horizontal smearing we're getting right now. Hopefully...